docs(testing): document evidence suite + eval scripts; rename to CapabilityUnavailableError
Answers a real user question: "how do I run these tests?" The answer
(make test-evidence) existed in the Makefile and in-line in the
reliability-harness doc but wasn't reachable from docs/testing.md,
which is the canonical entry point a contributor would look at.

Changes:
- docs/testing.md: new "Evidence suite — running the new tests in
  tests/evidence/" section with explicit commands for:
  * make test-evidence (deterministic suite, no network, ~30s)
  * make eval-summary-quality (Phase 9 nightly eval)
  * ./.venv/bin/python scripts/eval_llm_quality.py (Phase 10 reliability harness
    with multi-sample variance + judge-LLM grading)
  * ./.venv/bin/python scripts/record_golden.py (optional real-API re-recording)
- README.md: update the docs table to mention the new commands.

Lint + format pass applied across the evidence suite:
- metrics.py: drop redundant int(round(...)) per RUF046.
- llm/base.py + gemini_search.py + test_llm_search_routing.py:
  rename CapabilityUnavailable → CapabilityUnavailableError per N818.
- test_enrichment_services.py + test_fetch_services.py: annotate
  class-level mutable defaults with ClassVar per RUF012.
- test_eval_harness.py: en-dash → hyphen per RUF002.
- test_stats_services.py: itertools.pairwise() + iterable unpacking
  per RUF007 / RUF005.
- test_synthesis_services.py: remove ambiguous ∪ unicode from comment
  per RUF003.
- scripts/eval_llm_quality.py + record_golden.py: ./.venv/bin/ruff import sort
  (I-lint fixes).
- ./.venv/bin/ruff format pass touched 23 files for line wrapping / whitespace.

No behaviour change. 1497 tests pass; 100% branch ./.venv/bin/coverage preserved.

+++ social_research_probe/evals/metrics.py
@@ -9,7 +9,6 @@ from __future__ import annotations
 import re
 import statistics
 
-
 _PROPER_NOUN_RE = re.compile(r"\b[A-Z][A-Za-z0-9]{2,}\b")
 
 
@@ -22,9 +21,7 @@ def coverage(summary: str, required_tokens: list[str]) -> float:
     return hits / len(required_tokens)
 
 
-def hallucinated_names(
-    summary: str, source_text: str, allowed: list[str]
-) -> list[str]:
+def hallucinated_names(summary: str, source_text: str, allowed: list[str]) -> list[str]:
     """Return sorted proper nouns present in summary but not in source/allowed.
 
     Case-insensitive comparison. ``source_text`` is typically the transcript
@@ -34,11 +31,7 @@ def hallucinated_names(
     source_lower = source_text.lower()
     candidates = set(_PROPER_NOUN_RE.findall(summary))
     return sorted(
-        {
-            w
-            for w in candidates
-            if w.lower() not in allowed_lower and w.lower() not in source_lower
-        }
+        {w for w in candidates if w.lower() not in allowed_lower and w.lower() not in source_lower}
     )
 
 
@@ -65,6 +58,6 @@ def aggregate(values: list[float]) -> dict[str, float]:
         "stdev": statistics.stdev(values),
         "min": ordered[0],
         "max": ordered[-1],
-        "p5": ordered[max(0, int(round(0.05 * (n - 1))))],
-        "p95": ordered[min(n - 1, int(round(0.95 * (n - 1))))],
+        "p5": ordered[max(0, round(0.05 * (n - 1)))],
+        "p95": ordered[min(n - 1, round(0.95 * (n - 1)))],
     }
